+++ racetrack_commons/racetrack_commons/entities/plugin_client.py
@@ -14,6 +14,9 @@ class LifecyclePluginClient:
         response = self.lc_client.request_list('get', '/api/v1/plugins')
         return parse_dict_datamodels(response, PluginManifest)
 
+    def get_job_type_versions(self) -> List[str]:
+        return self.lc_client.request_list('get', '/api/v1/job_type/versions')
+
     def get_plugin_docs(self, plugin_name: str) -> Optional[str]:
         return self.lc_client.request_dict('get', f'/api/v1/plugin/{plugin_name}/docs')